seoul: extend for use with base-hw
Set some Vcpu_fields to charged in the case of pause exit to keep the
contrib code happy, because on base-hw pause exits are asynchronous
signals without state transfrer.

Reflect the state that is resumed from back to base-hw. For the case of
an outstanding pause request, this allows for distinction between a
resume from a previous dispatch (before the pause request is handled)
and the resume from handling the actual pause request itself.

Ref genodelabs/genode#4826